+++ NEWS.md
@@ -3,6 +3,16 @@
 Version numbers follow [Semantic Versioning](https://semver.org/).
 
 
+# portalcasting 0.35.0
+*In Progress*
+
+### Updating model controls
+* `time`
+* `response` with `type` and `link` 
+
+### Developing evaluate
+* `evaluate_casts`
+
 # [portalcasting 0.34.0](https://github.com/weecology/portalcasting/releases/tag/v0.34.0)
 *2022-04-07*